requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.2
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
executing==2.2.0
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.5.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
twilio==9.6.3
typing_extensions==4.14.0
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.1
